+++ src/raycast.c
@@ -6,32 +6,12 @@
 /*   By: amagnell <amagnell@student.42barcelon      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/30 11:20:16 by amagnell          #+#    #+#             */
-/*   Updated: 2025/01/30 11:20:17 by amagnell         ###   ########.fr       */
+/*   Updated: 2025/01/30 11:34:49 by amagnell         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "raycast.h"
 #include "MLX42/MLX42.h"
-#include <stdio.h>
-#include "texture.h"
-
-void	render_ray(int x, t_raycast r, t_cub *cub)
-{
-	int		y;
-
-	y = 0;
-	while (y < r.wall_start)
-	{
-		mlx_put_pixel(cub->img, x, y, cub->ass.c_color);
-		y++;
-	}
-	y = render_texture(x, y, r, cub);
-	while (y < HEIGHT)
-	{
-		mlx_put_pixel(cub->img, x, y, cub->ass.f_color);
-		y++;
-	}
-}
 
 int	calc_wall_height(t_raycast *r, t_cub *cub)
 {
